refactor MegaMoEV2 intra and inter-node backends
