+++ aiter/configs/model_configs/kimik3_a4w4_tuned_fmoe.csv
@@ -16,3 +16,18 @@ gfx950,256,4096,3584,384,896,16,ActivationType.Situv2,torch.bfloat16,torch.float
 gfx950,256,8192,3584,384,896,16,ActivationType.Situv2,torch.bfloat16,torch.float4_e2m1fn_x2,torch.float4_e2m1fn_x2,QuantType.per_1x32,1,0,64,0,422.9114,flydsl_moe1_afp4_wfp4_bf16_t64x128x256_w4_bnt0_xcd4_fp4,0.1%,597.3431,flydsl_moe2_layout_afp4_wfp4_bf16_t64x256x128_reduce_sbm64_sp1601,0.0%,1020.2545,0,0,0,1060.84,3712.27,
 gfx950,256,16384,3584,384,896,16,ActivationType.Situv2,torch.bfloat16,torch.float4_e2m1fn_x2,torch.float4_e2m1fn_x2,QuantType.per_1x32,1,0,64,0,642.4294,flydsl_moe1_afp4_wfp4_bf16_t64x128x256_w3_bnt0_xcd4_fp4,0.1%,1074.0772,flydsl_moe2_layout_afp4_wfp4_bf16_t64x256x128_reduce_sbm64_sp801,0.0%,1716.5066,0,0,0,1261.09,2257.8,
 gfx950,256,32768,3584,384,896,16,ActivationType.Situv2,torch.bfloat16,torch.float4_e2m1fn_x2,torch.float4_e2m1fn_x2,QuantType.per_1x32,1,0,64,0,1111.7422,flydsl_moe1_afp4_wfp4_bf16_t64x128x256_w2_bnt0_xcd4_fp4,0.1%,2185.1161,flydsl_moe2_layout_afp4_wfp4_bf16_t64x256x128_reduce_sbm64_sp1601,0.0%,3296.8583,0,0,0,1313.17,1228.96,
+gfx950,256,1,3584,3072,56,16,ActivationType.Situv2,torch.bfloat16,torch.float4_e2m1fn_x2,torch.float4_e2m1fn_x2,QuantType.per_1x32,1,0,16,0,,flydsl_mxmoe_g1_a4w4_16x256x256_f16in_nt,,,flydsl_moe2_layout_afp4_wfp4_bf16_t16x256x128_atomic_sbm16
+gfx950,256,2,3584,3072,56,16,ActivationType.Situv2,torch.bfloat16,torch.float4_e2m1fn_x2,torch.float4_e2m1fn_x2,QuantType.per_1x32,1,0,16,0,,flydsl_mxmoe_g1_a4w4_16x256x256_f16in_nt,,,flydsl_moe2_layout_afp4_wfp4_bf16_t32x128x256_reduce_sbm32
+gfx950,256,3,3584,3072,56,16,ActivationType.Situv2,torch.bfloat16,torch.float4_e2m1fn_x2,torch.float4_e2m1fn_x2,QuantType.per_1x32,1,0,16,0,,flydsl_mxmoe_g1_a4w4_16x256x256_f16in_nt,,,flydsl_moe2_layout_afp4_wfp4_bf16_t16x128x128_atomic_sbm16
+gfx950,256,4,3584,3072,56,16,ActivationType.Situv2,torch.bfloat16,torch.float4_e2m1fn_x2,torch.float4_e2m1fn_x2,QuantType.per_1x32,1,0,16,0,,flydsl_mxmoe_g1_a4w4_16x256x256_f16in_nt,,,flydsl_moe2_layout_afp4_wfp4_bf16_t16x128x256_atomic_persist_nt_sbm16
+gfx950,256,8,3584,3072,56,16,ActivationType.Situv2,torch.bfloat16,torch.float4_e2m1fn_x2,torch.float4_e2m1fn_x2,QuantType.per_1x32,1,0,32,0,,flydsl_mxmoe_g1_a4w4_16x256x256_f16in_nt,,,flydsl_moe2_layout_afp4_wfp4_bf16_t32x128x256_atomic_persist_sbm32
+gfx950,256,16,3584,3072,56,16,ActivationType.Situv2,torch.bfloat16,torch.float4_e2m1fn_x2,torch.float4_e2m1fn_x2,QuantType.per_1x32,1,0,32,0,,flydsl_mxmoe_g1_a4w4_16x256x256_f16in_nt,,,flydsl_moe2_layout_afp4_wfp4_bf16_t16x128x128_atomic_nt_sbm16
+gfx950,256,32,3584,3072,56,16,ActivationType.Situv2,torch.bfloat16,torch.float4_e2m1fn_x2,torch.float4_e2m1fn_x2,QuantType.per_1x32,1,0,32,0,,flydsl_mxmoe_g1_a4w4_16x256x256_f16in_nt,,,flydsl_moe2_layout_afp4_wfp4_bf16_t32x128x128_atomic_persist_sbm32
+gfx950,256,64,3584,3072,56,16,ActivationType.Situv2,torch.bfloat16,torch.float4_e2m1fn_x2,torch.float4_e2m1fn_x2,QuantType.per_1x32,1,0,32,0,,flydsl_mxmoe_g1_a4w4_16x256x256_f16in_nt,,,flydsl_moe2_layout_afp4_wfp4_bf16_t32x128x256_atomic_sbm32
+gfx950,256,128,3584,3072,56,16,ActivationType.Situv2,torch.bfloat16,torch.float4_e2m1fn_x2,torch.float4_e2m1fn_x2,QuantType.per_1x32,1,0,32,0,,flydsl_mxmoe_g1_a4w4_16x256x256_f16in_nt,,,flydsl_moe2_layout_afp4_wfp4_bf16_t64x256x128_atomic_nt_sbm64
+gfx950,256,256,3584,3072,56,16,ActivationType.Situv2,torch.bfloat16,torch.float4_e2m1fn_x2,torch.float4_e2m1fn_x2,QuantType.per_1x32,1,0,32,0,,flydsl_mxmoe_g1_a4w4_16x256x256_f16in_nt,,,flydsl_moe2_layout_afp4_wfp4_bf16_t32x128x128_atomic_persist_nt_sbm32
+gfx950,256,512,3584,3072,56,16,ActivationType.Situv2,torch.bfloat16,torch.float4_e2m1fn_x2,torch.float4_e2m1fn_x2,QuantType.per_1x32,1,0,64,0,,flydsl_mxmoe_g1_a4w4_16x256x256_f16in_nt,,,flydsl_moe2_layout_afp4_wfp4_bf16_t64x256x128_atomic_nt_sbm64
+gfx950,256,1024,3584,3072,56,16,ActivationType.Situv2,torch.bfloat16,torch.float4_e2m1fn_x2,torch.float4_e2m1fn_x2,QuantType.per_1x32,1,0,64,0,,flydsl_mxmoe_g1_a4w4_16x256x256_f16in_nt,,,flydsl_moe2_layout_afp4_wfp4_bf16_t128x256x256_atomic_nt_sbm128
+gfx950,256,2048,3584,3072,56,15,ActivationType.Situv2,torch.bfloat16,torch.float4_e2m1fn_x2,torch.float4_e2m1fn_x2,QuantType.per_1x32,1,0,64,0,,flydsl_mxmoe_g1_a4w4_128x256x256,,,flydsl_moe2_layout_afp4_wfp4_bf16_t64x256x128_atomic_persist_sbm64,,,0,0,0,0,0,
+gfx950,256,4096,3584,3072,56,16,ActivationType.Situv2,torch.bfloat16,torch.float4_e2m1fn_x2,torch.float4_e2m1fn_x2,QuantType.per_1x32,1,0,128,0,,flydsl_mxmoe_g1_a4w4_128x256x256,,,flydsl_moe2_layout_afp4_wfp4_bf16_t64x256x128_atomic_persist_sbm64
+gfx950,256,8192,3584,3072,56,16,ActivationType.Situv2,torch.bfloat16,torch.float4_e2m1fn_x2,torch.float4_e2m1fn_x2,QuantType.per_1x32,1,0,128,0,,flydsl_mxmoe_g1_a4w4_128x256x256,,,flydsl_moe2_layout_afp4_wfp4_bf16_t128x128x128_atomic_sbm128
