+++ trace_data/wrapper_ep16_a4w4_20260903/wrapper_perf_20260903.jsonl
@@ -1,0 +1,5 @@
+{"category": "ep16_a4w4_moe", "metrics": {"active_experts_max": 56.0, "active_experts_mean": 56.0, "combine_avg_ms": 0.274, "combine_best_ms": 0.0976, "combine_gbps_best": 19.16, "combine_gbps_mean": 7.26, "combine_gbps_worst": 6.3, "combine_worst_ms": 0.2931, "dispatch_avg_ms": 0.0687, "dispatch_best_ms": 0.0606, "dispatch_gbps_best": 7.58, "dispatch_gbps_mean": 6.74, "dispatch_gbps_worst": 6.03, "dispatch_worst_ms": 0.0761, "local_hits_max": 256.0, "local_hits_mean": 256.0, "moe_avg_ms": 0.2538, "moe_best_ms": 0.2388, "moe_gbps_best": 3902.19, "moe_gbps_mean": 3670.71, "moe_gbps_worst": 2201.63, "moe_rel_l2": null, "moe_tflops_best": 70.83, "moe_tflops_mean": 66.63, "moe_tflops_worst": 39.96, "moe_worst_ms": 0.4232, "total_recv": 256.0}, "params": {"bs": 16, "experts": 896, "inter_dim": 3072, "local_experts": 56, "model_dim": 3584, "quant_type": "per_1x32_a4w4", "topk": 16, "world_size": 16}, "stat_iters": 20, "total_iters": 100, "ts": 1788423586.0050697}
+{"category": "ep16_a4w4_moe", "metrics": {"active_experts_max": 56.0, "active_experts_mean": 56.0, "combine_avg_ms": 0.2484, "combine_best_ms": 0.0927, "combine_gbps_best": 39.6, "combine_gbps_mean": 15.83, "combine_gbps_worst": 13.74, "combine_worst_ms": 0.2681, "dispatch_avg_ms": 0.0699, "dispatch_best_ms": 0.0601, "dispatch_gbps_best": 15.3, "dispatch_gbps_mean": 13.33, "dispatch_gbps_worst": 11.6, "dispatch_worst_ms": 0.0791, "local_hits_max": 512.0, "local_hits_mean": 512.0, "moe_avg_ms": 0.2666, "moe_best_ms": 0.251, "moe_gbps_best": 3738.78, "moe_gbps_mean": 3520.01, "moe_gbps_worst": 2270.73, "moe_rel_l2": null, "moe_tflops_best": 134.75, "moe_tflops_mean": 126.86, "moe_tflops_worst": 81.84, "moe_worst_ms": 0.4133, "total_recv": 512.0}, "params": {"bs": 32, "experts": 896, "inter_dim": 3072, "local_experts": 56, "model_dim": 3584, "quant_type": "per_1x32_a4w4", "topk": 16, "world_size": 16}, "stat_iters": 20, "total_iters": 100, "ts": 1788423586.1120267}
+{"category": "ep16_a4w4_moe", "metrics": {"active_experts_max": 56.0, "active_experts_mean": 56.0, "combine_avg_ms": 0.2128, "combine_best_ms": 0.0956, "combine_gbps_best": 76.82, "combine_gbps_mean": 36.49, "combine_gbps_worst": 30.53, "combine_worst_ms": 0.2422, "dispatch_avg_ms": 0.0913, "dispatch_best_ms": 0.0713, "dispatch_gbps_best": 25.77, "dispatch_gbps_mean": 21.0, "dispatch_gbps_worst": 16.53, "dispatch_worst_ms": 0.111, "local_hits_max": 1024.0, "local_hits_mean": 1024.0, "moe_avg_ms": 0.2777, "moe_best_ms": 0.2634, "moe_gbps_best": 3614.61, "moe_gbps_mean": 3429.06, "moe_gbps_worst": 2538.12, "moe_rel_l2": null, "moe_tflops_best": 256.81, "moe_tflops_mean": 243.63, "moe_tflops_worst": 180.33, "moe_worst_ms": 0.3751, "total_recv": 1024.0}, "params": {"bs": 64, "experts": 896, "inter_dim": 3072, "local_experts": 56, "model_dim": 3584, "quant_type": "per_1x32_a4w4", "topk": 16, "world_size": 16}, "stat_iters": 20, "total_iters": 100, "ts": 1788423586.2121546}
+{"category": "ep16_a4w4_moe", "metrics": {"active_experts_max": 56.0, "active_experts_mean": 56.0, "combine_avg_ms": 0.1984, "combine_best_ms": 0.12, "combine_gbps_best": 91.75, "combine_gbps_mean": 57.04, "combine_gbps_worst": 50.95, "combine_worst_ms": 0.218, "dispatch_avg_ms": 0.0927, "dispatch_best_ms": 0.0815, "dispatch_gbps_best": 33.78, "dispatch_gbps_mean": 30.06, "dispatch_gbps_worst": 26.62, "dispatch_worst_ms": 0.1035, "local_hits_max": 1536.0, "local_hits_mean": 1536.0, "moe_avg_ms": 0.298, "moe_best_ms": 0.2814, "moe_gbps_best": 3432.09, "moe_gbps_mean": 3240.7, "moe_gbps_worst": 2632.08, "moe_rel_l2": null, "moe_tflops_best": 360.6, "moe_tflops_mean": 340.49, "moe_tflops_worst": 276.55, "moe_worst_ms": 0.3669, "total_recv": 1536.0}, "params": {"bs": 96, "experts": 896, "inter_dim": 3072, "local_experts": 56, "model_dim": 3584, "quant_type": "per_1x32_a4w4", "topk": 16, "world_size": 16}, "stat_iters": 20, "total_iters": 100, "ts": 1788423586.3123758}
+{"category": "ep16_a4w4_moe", "metrics": {"active_experts_max": 56.0, "active_experts_mean": 56.0, "combine_avg_ms": 0.188, "combine_best_ms": 0.1095, "combine_gbps_best": 134.08, "combine_gbps_mean": 80.9, "combine_gbps_worst": 68.23, "combine_worst_ms": 0.2166, "dispatch_avg_ms": 0.1003, "dispatch_best_ms": 0.0777, "dispatch_gbps_best": 47.28, "dispatch_gbps_mean": 38.29, "dispatch_gbps_worst": 30.14, "dispatch_worst_ms": 0.1219, "local_hits_max": 2048.0, "local_hits_mean": 2048.0, "moe_avg_ms": 0.313, "moe_best_ms": 0.295, "moe_gbps_best": 3319.6, "moe_gbps_mean": 3129.44, "moe_gbps_worst": 2647.44, "moe_rel_l2": null, "moe_tflops_best": 458.57, "moe_tflops_mean": 432.3, "moe_tflops_worst": 365.72, "moe_worst_ms": 0.3699, "total_recv": 2048.0}, "params": {"bs": 128, "experts": 896, "inter_dim": 3072, "local_experts": 56, "model_dim": 3584, "quant_type": "per_1x32_a4w4", "topk": 16, "world_size": 16}, "stat_iters": 20, "total_iters": 100, "ts": 1788423586.4150434}
